+++ src/test-utils.tsx
@@ -12,7 +12,7 @@ import {
   DocumentNode,
 } from 'graphql';
 
-import { print } from 'graphql-tag/bundledPrinter';
+import { print } from 'graphql';
 
 
 import ApolloProvider from './ApolloProvider';
@@ -30,7 +30,7 @@ export class MockedProvider extends React.Component<any, any> {
 
   render() {
     return (
-      <ApolloProvider client={this.client || this.props.client}>
+      <ApolloProvider client={this.client || this.props.client} store={this.props.store || null}>
         {this.props.children}
       </ApolloProvider>
     );
@@ -107,11 +107,11 @@ export class MockNetworkInterface implements NetworkInterface {
       if (!mockedResponse.result && !mockedResponse.error) {
         throw new Error('Mocked response should contain either result or error.');
       }
-      this.addMockedReponse(mockedResponse);
+      this.addMockedResponse(mockedResponse);
     });
   }
 
-  public addMockedReponse(mockedResponse: MockedResponse) {
+  public addMockedResponse(mockedResponse: MockedResponse) {
     const key = requestToKey(mockedResponse.request);
     let mockedResponses = this.mockedResponsesByKey[key];
     if (!mockedResponses) {
@@ -204,7 +204,7 @@ export class MockSubscriptionNetworkInterface extends MockNetworkInterface imple
       throw new Error('Network interface does not have subscription associated with this request.');
     }
 
-  };
+  }
 
   public fireResult(id: number) {
     const handler = this.handlersById[id];