+++ test/react-web/client/graphql/queries.test.tsx
@@ -1,12 +1,13 @@
 /// <reference types="jest" />
 
 import * as React from 'react';
+import * as PropTypes from 'prop-types';
 import * as ReactDOM from 'react-dom';
 import * as renderer from 'react-test-renderer';
 import { mount } from 'enzyme';
 import gql from 'graphql-tag';
-
-import ApolloClient, { ApolloError } from 'apollo-client';
+import ApolloClient, { ApolloError, ObservableQuery } from 'apollo-client';
+import { NetworkInterface } from 'apollo-client/transport/networkInterface';
 import { connect } from 'react-redux';
 import { withState } from 'recompose';
 
@@ -1156,7 +1157,7 @@ describe('queries', () => {
             variables: { skip: 2 },
             updateQuery: (prev, { fetchMoreResult }) => ({
               allPeople: {
-                people: prev.allPeople.people.concat(fetchMoreResult.data.allPeople.people),
+                people: prev.allPeople.people.concat(fetchMoreResult.allPeople.people),
               },
             }),
           }).then(wrap(done, result => {
@@ -1603,7 +1604,7 @@ describe('queries', () => {
     }
 
     (ContextContainer as any).childContextTypes = {
-      color: React.PropTypes.string,
+      color: PropTypes.string,
     };
 
     let count = 0;
@@ -1622,7 +1623,7 @@ describe('queries', () => {
     }
 
     (ChildContextContainer as any).contextTypes = {
-      color: React.PropTypes.string,
+      color: PropTypes.string,
     };
 
     renderer.create(
@@ -1805,7 +1806,7 @@ describe('queries', () => {
           getMorePeople: () => fetchMore({
             variables: { cursor },
             updateQuery(prev, { fetchMoreResult }) {
-              const { data: { allPeople: { cursor, people } } } = fetchMoreResult;
+              const { allPeople: { cursor, people } } = fetchMoreResult;
               return {
                 allPeople: {
                   cursor,
@@ -2157,7 +2158,9 @@ describe('queries', () => {
     );
 
     expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
-    const queryObservable1 = (client as any).queryManager.observableQueries['1'].observableQuery;
+    const queryObservable1: ObservableQuery<any> = (client as any).queryManager.observableQueries['1'].observableQuery;
+
+    const originalOptions = Object.assign({}, queryObservable1.options);
 
     wrapper1.unmount();
 
@@ -2170,13 +2173,86 @@ describe('queries', () => {
     );
 
     expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
-    const queryObservable2 = (client as any).queryManager.observableQueries['1'].observableQuery;
+    const queryObservable2: ObservableQuery<any> = (client as any).queryManager.observableQueries['1'].observableQuery;
+
+    const recycledOptions = queryObservable2.options;
 
     expect(queryObservable1).toBe(queryObservable2);
+    expect(recycledOptions).toEqual(originalOptions);
 
     wrapper2.unmount();
 
     expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
+  });
+
+  it('will not try to refetch recycled `ObservableQuery`s when resetting the client store', () => {
+    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
+    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
+    const networkInterface = {
+      query: jest.fn(),
+    } as NetworkInterface;
+    const client = new ApolloClient({ networkInterface, addTypename: false });
+
+    @graphql(query)
+    class Container extends React.Component<any, any> {
+      render () {
+        return null;
+      }
+    }
+
+    const wrapper1 = renderer.create(
+      <ApolloProvider client={client}>
+        <Container/>
+      </ApolloProvider>
+    );
+
+    expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
+    const queryObservable1 = (client as any).queryManager.observableQueries['1'].observableQuery;
+
+    // The query should only have been invoked when first mounting and not when resetting store
+    expect(networkInterface.query).toHaveBeenCalledTimes(1);
+
+    wrapper1.unmount();
+
+    expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
+    const queryObservable2 = (client as any).queryManager.observableQueries['1'].observableQuery;
+
+    expect(queryObservable1).toBe(queryObservable2);
+
+    client.resetStore();
+
+    // The query should not have been fetch again
+    expect(networkInterface.query).toHaveBeenCalledTimes(1);
+  });
+
+  it('will refetch active `ObservableQuery`s when resetting the client store', () => {
+    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
+    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
+    const networkInterface = {
+      query: jest.fn(),
+    } as NetworkInterface;
+    const client = new ApolloClient({ networkInterface, addTypename: false });
+
+    @graphql(query)
+    class Container extends React.Component<any, any> {
+      render () {
+        return null;
+      }
+    }
+
+    const wrapper1 = renderer.create(
+      <ApolloProvider client={client}>
+        <Container/>
+      </ApolloProvider>
+    );
+
+    expect(Object.keys((client as any).queryManager.observableQueries)).toEqual(['1']);
+
+    expect(networkInterface.query).toHaveBeenCalledTimes(1);
+
+    client.resetStore();
+
+    expect(networkInterface.query).toHaveBeenCalledTimes(2);
   });
 
   it('will recycle `ObservableQuery`s when re-rendering a portion of the tree', done => {
@@ -2456,7 +2532,7 @@ describe('queries', () => {
       }
     }
 
-    @graphql(query)
+    @graphql(query, { options: { notifyOnNetworkStatusChange: true } })
     class Query extends React.Component<any, any> {
       componentDidMount() {
         refetchQuery = () => this.props.data.refetch();
@@ -2503,5 +2579,45 @@ describe('queries', () => {
       { loading: false, a: 7, b: 8, c: 9 },
     ]);
   });
+
+  it('passes any cached data when there is a GraphQL error', (done) => {
+    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
+    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
+    const networkInterface = mockNetworkInterface(
+      { request: { query }, result: { data } },
+      { request: { query }, error: new Error('No Network Connection') }
+    );
+    const client = new ApolloClient({ networkInterface, addTypename: false });
+
+    let count = 0;
+    @graphql(query, { options: { notifyOnNetworkStatusChange: true } })
+    class Container extends React.Component<any, any> {
+      componentWillReceiveProps = (props) => {
+        switch (count++) {
+          case 0:
+            expect(props.data.allPeople).toEqual(data.allPeople);
+            props.data.refetch();
+            break;
+          case 1:
+            expect(props.data.loading).toBe(true);
+            expect(props.data.allPeople).toEqual(data.allPeople);
+            break;
+          case 2:
+            expect(props.data.loading).toBe(false);
+            expect(props.data.error).toBeTruthy();
+            expect(props.data.allPeople).toEqual(data.allPeople);
+            done();
+            break;
+        }
+      }
+
+      render() {
+        return null;
+      }
+    };
+
+    const wrapper = renderer.create(<ApolloProvider client={client}><Container /></ApolloProvider>);
+  }));
+
 
 });